Crate vaultist

Modules

bitwarden_reader

chrome_reader

firefox_reader

keychain_reader

password_generator

Structs

Entry

An entry in the vault

Secret

An encrypted entry held within the vault

Vault

A vault for storing secrets

Enums

VaultVersion

The possible versions of a Vaultist vault

Functions

decrypt_bytes

Decrypt a byte array using the vault’s key.

decrypt_secret

Decrypt a secret from the vault.

encrypt_bytes

Encrypt a byte array using a vault’s key.

generate_nonce

Generates a random nonce for use with the XChaCha20Poly1305 cipher.

generate_salt

Generates a random salt for use with the Argon2 algorithm.