Module vaultist::bitwarden_reader

Structs

BitwardenCard

A card Bitwarden record

BitwardenField

A field in a Bitwarden record

BitwardenFolder

A folder in a Bitwarden vault

BitwardenIdentity

An identity Bitwarden record

BitwardenLogin

A login Bitwarden record

BitwardenRecordCSV

A record in a Bitwarden vault exported in CSV format

BitwardenRecordJSON

A record in a Bitwarden vault exported in JSON format

BitwardenSecureNote

A secure note Bitwarden record

BitwardenURI

A URI in a Bitwarden login record

BitwardenVault

An exported Bitwarden vault

Functions

get_secrets_from_bitwarden

Gets a list of secrets from an exported Bitwarden vault.

get_secrets_from_bitwarden_csv

Gets a list of secrets from an exported Bitwarden vault in CSV format.

get_secrets_from_bitwarden_json

Gets a list of secrets from an exported Bitwarden vault in JSON format.