Function vaultist::password_generator::correct_horse_battery_staple

pub fn correct_horse_battery_staple(
    count: usize,
    length: usize,
    numbers: bool,
    lowercase_letters: bool,
    uppercase_letters: bool,
    symbols: bool,
    spaces: bool,
    exclude_similar_characters: bool
) -> Vec<String> 

Generate a CorrectHorseBatteryStaple-style password.
Optionally, can use two additional elements:

A randomly-picked special character to separate each element (note: the separator will always be a space if spaces is true)

A randomly-picked one- or two-digit number (note: if numbers are allowed, the password will always end with one)

Arguments

count - The number of passwords to generate.

length - The length of the generated passwords.

numbers - Passwords are should contain at least one number.

lowercase_letters - Passwords should contain at least one lowercase letter.

uppercase_letters - Passwords should contain at least one uppercase letter.

symbols - Passwords should contain at least one special character.

spaces - Passwords should use a space to separate each password element.

exclude_similar_characters - Whether or not to exclude similar looking ASCII characters (iI1loO0"'`|).