Struct vaultist::BitwardenRecordJSON

pub struct BitwardenRecordJSON {

    pub id: String,
    pub organization_id: Option<String>,
    pub folder_id: Option<String>,
    pub type_: usize,
    pub reprompt: Option<usize>,
    pub name: String,
    pub notes: Option<String>,
    pub favorite: bool,
    pub fields: Option<Vec<BitwardenField>>,
    pub login: Option<BitwardenLogin>,
    pub secure_note: Option<BitwardenSecureNote>,
    pub card: Option<BitwardenCard>,
    pub identity: Option<BitwardenIdentity>,
    pub collection_ids: Option<Vec<String>>,

}

A record in a Bitwarden vault exported in JSON format

Fields
id: String

The ID of the record

organization_id: Option<String>

The ID of the organisation the record is in

folder_id: Option<String>

The ID of the folder the record is in

type_: usize

The type of the record (see: https://bitwarden.com/help/managing-items/).

Can be:

Login (1)

Secure Note (2)

Card (3)

Identity (4)

See: https://bitwarden.com/help/cli/#enums

reprompt: Option<usize>

Whether or not the vault password should be reprompted to view the record

name: String

The name of the record

notes: Option<String>

The notes associated with the record

favorite: bool

Whether or not the record is a favourite

fields: Option<Vec<BitwardenField>>

The fields associated with the record

login: Option<BitwardenLogin>

The login information associated with the record

secure_note: Option<BitwardenSecureNote>

The secure note information associated with the record

card: Option<BitwardenCard>

The card information associated with the record

identity: Option<BitwardenIdentity>

The identity information associated with the record

collection_ids: Option<Vec<String>>

The IDs of the collections the record is in

Implementations

impl BitwardenRecordJSON

pub fn new(
id: String,
organization_id: Option<String>,
folder_id: Option<String>,
type_: usize,
reprompt: Option<usize>,
name: String,
notes: Option<String>,
favorite: bool,
fields: Option<Vec<BitwardenField>>,
login: Option<BitwardenLogin>,
secure_note: Option<BitwardenSecureNote>,
card: Option<BitwardenCard>,
identity: Option<BitwardenIdentity>,
collection_ids: Option<Vec<String>>
) -> BitwardenRecordJSON