Crate vaultist

Structs

ELEMENTS

The possible non-separator elements of a CorrectHorseBatteryStaple-type password

Entry

An entry in the vault

SIMILAR_CHARACTERS

A vector of similar-looking characters

SYMBOLS

A vector of symbol characters

Secret

An encrypted entry held within the vault

Vault

A vault for storing secrets

Enums

CorrectHorseBatteryStapleElements

The possible elements of a CorrectHorseBatteryStaple-type password

Functions

correct_horse_battery_staple

Generate a CorrectHorseBatteryStaple-style password.
Optionally, can use two additional elements:

create_vault_from_password

Create a new vault from a user-supplied password.

decrypt_bytes

Decrypt a byte array using the vault’s key.

decrypt_secret

Decrypt a secret from the vault.

encrypt_bytes

Encrypt a byte array using a vault’s key.

generate_key_from_password_and_salt

Generate a vault’s key from a user-supplied password and a pre-generated salt.

generate_nonce

Generates a random nonce for use with the XChaCha20Poly1305 cipher.

get_password_feedback

Analyses whether or not a password is sufficiently resistant against an attack.